Constructive feedback from grader
Greetings, Ralph!

Congratulations on your good SIEMs Module 19 Challenge! You’ve brought your talents and knowledge to the table, and this is a superb report, Ralph.

Ralph, you’ve analyzed the outage correctly, and determined the time frame for system downtime and recovery after the DDOS. You’ve included a screenshot of your report including “_time”, “IP_ADDRESS”, “DOWNLOAD_MEGABITS”, “UPLOAD_MEGABITS”, and “ratio”. Well done on this deliverable!

In step 2, you’ve depicted the number of critical vulnerabilities with a Splunk report, and you’ve set a custom alert, but with the wrong parameters. You've made a good effort in this regard.

Lastly, Ralph, you’ve identified the time frame for the brute force attack, and established a baseline of 4 and threshold of 5 bad logins per hour for your custom alert. However, this threshold is far too low, and it will plague you with a false positive every hour. Try to examine the normal stretches of traffic for a better idea of your baseline.

Ralph, I wish you all the best for a bright future!

RAE
Central Grader , Jun 28, 2023 at 11:19pm